requirements.txt
bmp280==0.0.4
certifi==2022.9.24
charset-normalizer==2.1.1
i2cdevice==0.0.7
idna==3.4
influxdb-client==1.34.0
pimoroni-bme280==0.1.1
python-dateutil==2.8.2
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
typing_extensions==4.4.0
urllib3==1.26.12